Makefile:	Added UNIXWARE 2.x with shadow passwords from fja@extratech.com client.c:	Made prompt appear at debug level 0. Fixed strcasecmp redefinition.             Caused client to use set_blocking rather than making fcntl calls itself. dir.c:		Removed redundent snum parameters. includes.h:	Added SCO fixes. loadparm.c:	Made default 'files to hide' a null string. nmbd.c:		Removed O_NONBLOCK from pid file open for platforms that dont have it. proto.h:	Changed snum to cnum where needed. Changed is_xx_path to is_in_path             (now called via MACRO). quotas.c:	Swapped setuid/seteuid calls when restoring uid. reply.c:	Removed redundent snum parameters. server.c:	Changed snum to cnum where needed. Setup new veto_list, hide_list             namelists. Added standard_sub changes from             Stefaan A Eeckels <Stefaan.Eeckels@ecc.lu> and             Paul Rippin <pr3245@nopc.eurostat.cec.be> shmem.c:	Changed cast for sizeof to be int before negating. smb.h:		Added new veto_list, hide_list entries to connections.             Added IS_PRINT, IS_HIDDEN_PATH, IS_VETO_PATH macros. trans2.c:	Removed redundent snum parameters. util.c:		Added standard_sub_basic changes from             Stefaan A Eeckels <Stefaan.Eeckels@ecc.lu> and             Paul Rippin <pr3245@nopc.eurostat.cec.be>             Fixed up veto/hidden path processing so the paths are             pres-parsed and checked for wildcards (for speed). Jeremy (jallison@whistle.com)

+++ source/locking/shmem.c
@@ -167,7 +167,7 @@ static BOOL smb_shm_register_process(char *processreg_file, pid_t pid, BOOL *oth
    int smb_shm_processes_fd = -1;
    int nb_read;
    pid_t other_pid;
-   int seek_back = -sizeof(other_pid);
+   int seek_back = -((int)sizeof(other_pid));
    int free_slot = -1;
    int erased_slot;   
    
@@ -236,7 +236,7 @@ static BOOL smb_shm_unregister_process(char *processreg_file, pid_t pid)
    int smb_shm_processes_fd = -1;
    int nb_read;
    pid_t other_pid;
-   int seek_back = -sizeof(other_pid);
+   int seek_back = -((int)sizeof(other_pid));
    int erased_slot;
    BOOL found = False;
